+++ lol/Crypto/Lol.hs
@@ -7,8 +7,8 @@ module Crypto.Lol
 , module Crypto.Lol.LatticePrelude
 
 , module Crypto.Lol.Types.ZqBasic
-, module Crypto.Lol.Cyclotomic.Tensor.CTensor
-, module Crypto.Lol.Cyclotomic.Tensor.RepaTensor
+, module Crypto.Lol.Cyclotomic.Tensor.C
+, module Crypto.Lol.Cyclotomic.Tensor.Repa
 , module Crypto.Lol.Types.IrreducibleChar2) where
 
 import Crypto.Lol.Cyclotomic.Cyc
@@ -16,6 +16,6 @@ import Crypto.Lol.Gadget
 import Crypto.Lol.LatticePrelude
 
 import Crypto.Lol.Types.ZqBasic
-import Crypto.Lol.Cyclotomic.Tensor.CTensor
-import Crypto.Lol.Cyclotomic.Tensor.RepaTensor
+import Crypto.Lol.Cyclotomic.Tensor.C
+import Crypto.Lol.Cyclotomic.Tensor.Repa
 import Crypto.Lol.Types.IrreducibleChar2
