+++ lol/Crypto/Lol/Cyclotomic/Tensor/C/Extension.hs
@@ -4,7 +4,7 @@
 
 -- | CT-specific functions for embedding/twacing in various bases
 
-module Crypto.Lol.Cyclotomic.Tensor.CTensor.Extension
+module Crypto.Lol.Cyclotomic.Tensor.C.Extension
 ( embedPow', embedDec', embedCRT'
 , twacePowDec' -- , twaceCRT'
 , coeffs', powBasisPow'
