+++ lol/Crypto/Lol/Cyclotomic/Tensor/C/generalfuncs.c
