+++ lol/Crypto/Lol/Cyclotomic/Tensor/C/norm.c
