+++ lol/Crypto/Lol/Cyclotomic/Tensor/Repa/CRT.hs
@@ -3,7 +3,7 @@
 
 -- | Functions to support the chinese remainder transform on Repa arrays
 
-module Crypto.Lol.Cyclotomic.Tensor.RepaTensor.CRT
+module Crypto.Lol.Cyclotomic.Tensor.Repa.CRT
 ( scalarCRT'
 , fCRT, fCRTInv
 , mulGCRT', divGCRT'
@@ -12,7 +12,7 @@ module Crypto.Lol.Cyclotomic.Tensor.RepaTensor.CRT
 
 import Crypto.Lol.CRTrans
 import Crypto.Lol.Cyclotomic.Tensor
-import Crypto.Lol.Cyclotomic.Tensor.RepaTensor.RTCommon as RT
+import Crypto.Lol.Cyclotomic.Tensor.Repa.Common         as RT
 import Crypto.Lol.LatticePrelude                        as LP
 
 import Control.Applicative
