+++ lol/lol.cabal
@@ -30,7 +30,7 @@ extra-source-files:  README, CHANGES.md,
                      utils/TestTypes.hs,
                      utils/Utils.hs,
                      utils/Harness/Cyc.hs,
-                     Crypto/Lol/Cyclotomic/Tensor/CTensor/tensorTypes.h
+                     Crypto/Lol/Cyclotomic/Tensor/C/tensorTypes.h
 cabal-version:       >= 1.10
 description:
     Λ ○ λ (Lol) is a general-purpose library for ring-based lattice cryptography.
@@ -59,15 +59,17 @@ Flag opt
   Manual:      False
 
 library
-  Include-dirs: Crypto/Lol/Cyclotomic/Tensor/CTensor
-  C-sources: Crypto/Lol/Cyclotomic/Tensor/CTensor/basic.c,
-             Crypto/Lol/Cyclotomic/Tensor/CTensor/crt.c,
-             Crypto/Lol/Cyclotomic/Tensor/CTensor/g.c,
-             Crypto/Lol/Cyclotomic/Tensor/CTensor/generalfuncs.c,
-             Crypto/Lol/Cyclotomic/Tensor/CTensor/l.c,
-             Crypto/Lol/Cyclotomic/Tensor/CTensor/norm.c,
-             Crypto/Lol/Cyclotomic/Tensor/CTensor/random.c
-  Includes: Crypto/Lol/Cyclotomic/Tensor/CTensor/tensorTypes.h
+  Include-dirs: Crypto/Lol/Cyclotomic/Tensor/C
+  C-sources:    Crypto/Lol/Cyclotomic/Tensor/C/basic.c,
+                Crypto/Lol/Cyclotomic/Tensor/C/crt.c,
+                Crypto/Lol/Cyclotomic/Tensor/C/g.c,
+                Crypto/Lol/Cyclotomic/Tensor/C/generalfuncs.c,
+                Crypto/Lol/Cyclotomic/Tensor/C/l.c,
+                Crypto/Lol/Cyclotomic/Tensor/C/norm.c,
+                Crypto/Lol/Cyclotomic/Tensor/C/random.c
+
+  Includes:     Crypto/Lol/Cyclotomic/Tensor/C/tensorTypes.h
+
   default-language:   Haskell2010
 
   if flag(icc)
@@ -101,8 +103,8 @@ library
 
     Crypto.Lol.Cyclotomic.Tensor
     Crypto.Lol.Cyclotomic.Tensor.Representation
-    Crypto.Lol.Cyclotomic.Tensor.CTensor
-    Crypto.Lol.Cyclotomic.Tensor.RepaTensor
+    Crypto.Lol.Cyclotomic.Tensor.C
+    Crypto.Lol.Cyclotomic.Tensor.Repa
 
     Crypto.Lol.Types.Random
     Crypto.Lol.Types.FiniteField
@@ -115,71 +117,85 @@ library
     Crypto.Lol.Types.Complex
 
   other-modules:
-
     Crypto.Lol.FactoredDefs
     Crypto.Lol.PosBinDefs
     Crypto.Lol.GaussRandom
     Crypto.Lol.Types.ZmStar
     Crypto.Lol.Types.Numeric
-    Crypto.Lol.Cyclotomic.Tensor.RepaTensor.CRT
-    Crypto.Lol.Cyclotomic.Tensor.RepaTensor.Extension
-    Crypto.Lol.Cyclotomic.Tensor.RepaTensor.Dec
-    Crypto.Lol.Cyclotomic.Tensor.RepaTensor.GL
-    Crypto.Lol.Cyclotomic.Tensor.RepaTensor.RTCommon
-    Crypto.Lol.Cyclotomic.Tensor.CTensor.Extension
-    Crypto.Lol.Cyclotomic.Tensor.CTensor.Backend
+
+    -- Repa backend
+    Crypto.Lol.Cyclotomic.Tensor.Repa.CRT
+    Crypto.Lol.Cyclotomic.Tensor.Repa.Common
+    Crypto.Lol.Cyclotomic.Tensor.Repa.Dec
+    Crypto.Lol.Cyclotomic.Tensor.Repa.Extension
+    Crypto.Lol.Cyclotomic.Tensor.Repa.GL
+
+    -- C backend
+    Crypto.Lol.Cyclotomic.Tensor.C.Extension
+    Crypto.Lol.Cyclotomic.Tensor.C.Backend
 
   build-depends:
-    arithmoi >= 0.4.1.3 && <0.5,
-    base==4.8.*,
-    binary,
-    bytestring,
-    constraints,
-    containers >= 0.5.6.2 && < 0.6,
-    crypto-api,
-    data-default >= 0.3.0 && < 0.6,
-    deepseq >= 1.4.1.1 && <1.5,
-    MonadRandom >= 0.2 && < 0.5,
-    mtl >= 2.2.1 && < 2.3,
-    numeric-prelude >= 0.4.2 && < 0.5,
-    QuickCheck >= 2.8 && < 2.9,
-    random >= 1.1 && < 1.2,
-    reflection >= 1.5.1 && < 2.2,
-    repa==3.4.*,
-    singletons >= 1.1.2.1 && < 2.1,
-    storable-record >= 0.0.3 && < 0.1,
-    th-desugar >= 1.5.4 && < 1.6,
-    tagged-transformer >= 0.7 && < 0.9,
-    template-haskell  >=  2.2.0.0,
-    transformers >= 0.4.2.0 && < 0.5,
-    vector==0.11.*,
-    vector-th-unbox >= 0.2.1.0 && < 0.3
-
-  other-extensions: TemplateHaskell
+      base               == 4.8.*
+    , arithmoi           >= 0.4.1.3 && <0.5
+    , binary
+    , bytestring
+    , constraints
+    , containers         >= 0.5.6.2 && < 0.6
+    , crypto-api
+    , data-default       >= 0.3.0   && < 0.6
+    , deepseq            >= 1.4.1.1 && < 1.5
+    , MonadRandom        >= 0.2     && < 0.5
+    , mtl                >= 2.2.1   && < 2.3
+    , numeric-prelude    >= 0.4.2   && <0.5
+    , QuickCheck         >= 2.8     && < 2.9
+    , random             >= 1.1     && < 1.2
+    , reflection         >= 1.5.1   && < 2.2
+    , repa               == 3.4.*
+    , singletons         >= 1.1.2.1 && < 2.1
+    , storable-record    >= 0.0.3   && < 0.1
+    , tagged-transformer >= 0.7     && < 0.9
+    , template-haskell   >= 2.2.0.0
+    , th-desugar         >= 1.5.4   && < 1.6
+    , transformers       >= 0.4.2.0 && < 0.5
+    , vector             == 0.11.*
+    , vector-th-unbox    >= 0.2.1.0 && < 0.3
 
 test-suite test-lol
   type:               exitcode-stdio-1.0
   hs-source-dirs:     tests,utils
   default-language:   Haskell2010
   main-is:            Main.hs
+  other-modules:
+    -- utils
+    Apply
+    Gen
+    Harness.Cyc
+    Tests
+    Utils
+
+    -- tests
+    CycTests
+    TensorTests
+    TestTypes
+    ZqTests
 
   ghc-options: -threaded -rtsopts
 
   build-depends:
-    base,
-    constraints,
-    deepseq,
-    DRBG,
-    lol,
-    MonadRandom,
-    mtl,
-    QuickCheck >= 2.8 && < 2.9,
-    random,
-    repa,
-    singletons,
-    test-framework >= 0.8 && < 0.9,
-    test-framework-quickcheck2 >= 0.3 && < 0.4,
-    vector
+      base
+    , constraints
+    , deepseq
+    , DRBG
+    , lol                        == 0.3.*
+    , MonadRandom
+    , mtl
+    , QuickCheck                 >= 2.8 && < 2.9
+    , random
+    , repa
+    , singletons
+    , test-framework             >= 0.8 && < 0.9
+    , test-framework-quickcheck2 >= 0.3 && < 0.4
+    , vector
 
 Benchmark bench-lol
   type:               exitcode-stdio-1.0
@@ -194,14 +210,17 @@ Benchmark bench-lol
   -- ghc-options: -fno-liberate-case -funfolding-use-threshold1000 -funfolding-keeness-factor1000
 
   build-depends:
-    base,
-    criterion,
-    deepseq,
-    DRBG,
-    lol,
-    MonadRandom,
-    mtl,
-    singletons,
-    transformers,
-    vector,
-    repa
+      base
+    , criterion
+    , deepseq
+    , DRBG
+    , lol
+    , MonadRandom
+    , mtl
+    , repa
+    , singletons
+    , transformers
+    , vector
+
+-- vim: nospell
+
